requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
dj-database-url==0.5.0
Django==2.1
django-cors-headers==3.4.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
psycopg2==2.8.5
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==5.1.0
